Move ClockWater to ignite-catalog
